refactor/#38: 대댓글 swagger 부족한 부분 추가 및 dto 추가 및 수정
