+++ src/comments/comments.controller.ts
@@ -10,19 +10,21 @@ import {
   Delete,
 } from "@nestjs/common";
 import { CommentService } from "./comments.service";
-import { CreateCommentDto } from "./dtos/create-comment.dto";
-import { UpdateCommentDto } from "./dtos/update-comment.dto";
-import { GetCommentDto } from "./dtos/get-comment.dto";
+import { CreateCommentDto } from "./dtos/comment-dtos/create-comment.dto";
+import { UpdateCommentDto } from "./dtos/comment-dtos/update-comment.dto";
+import { GetCommentDto } from "./dtos/comment-dtos/get-comment.dto";
 import { ApiTags } from "@nestjs/swagger";
 import { ApiCreateComment } from "./swagger-decorators/comment-swagger/create-comment-decorator";
 import { ApiFindComments } from "./swagger-decorators/comment-swagger/find-comments-decorator";
 import { ApiUpdateComment } from "./swagger-decorators/comment-swagger/update-comment-decorator";
 import { ApiDeleteComment } from "./swagger-decorators/comment-swagger/delete-comments-decorator";
-import { GetReplyDto } from "./dtos/get_reply.dto";
+import { GetReplyDto } from "./dtos/replies-dtos/get_reply.dto";
 import { ApiFindRelies } from "./swagger-decorators/reply-swagger/find-reply-decorate";
 import { ApiDeleteReply } from "./swagger-decorators/reply-swagger/delete-reply-decorate";
 import { ApiUpdateReply } from "./swagger-decorators/reply-swagger/update-reply-decorate";
 import { ApiCreateReply } from "./swagger-decorators/reply-swagger/create-reply-decorater";
+import { CreateReplyDto } from "./dtos/replies-dtos/create-reply.dto";
+import { UpdateReplyDto } from "./dtos/replies-dtos/update-reply.dto";
 
 @Controller("comments")
 @ApiTags("Comments & Replies")
@@ -66,7 +68,7 @@ export class CommentController {
   @ApiCreateReply()
   createReply(
     @Param("commentNo", ParseIntPipe) commentNo: number,
-    @Body() content: CreateCommentDto,
+    @Body() content: CreateReplyDto,
   ) {
     const userNo = 1;
     return this.commentService.createReply(commentNo, userNo, content);
@@ -85,7 +87,7 @@ export class CommentController {
   @ApiUpdateReply()
   updateReply(
     @Param("replyNo", ParseIntPipe) replyNo: number,
-    @Body() content: UpdateCommentDto,
+    @Body() content: UpdateReplyDto,
   ) {
     return this.commentService.updateReply(replyNo, content);
   }
