+++ src/comments/dtos/replies-dtos/update-reply.dto.ts
@@ -1,0 +1,3 @@
+import { CreateReplyDto } from "./create-reply.dto";
+
+export class UpdateReplyDto extends CreateReplyDto {}
